Simplify function.
Sacrifice some efficiency for simpler code.
Most return values of function call are now ignored, just the profile is
important. Calculation of the first (zeroth) flux surface is moved
inside the for loop.

Only with this commit the comment added in the previous commit is
actually true (about return value beeing ignored).